Update basic-calculator page & move it to MyProject file directory
New basic-calculator have update feature (History of Equation & Numbers Trivia)

There is still a bug in the "Numbers Trivia" when changing the screen size, it is recommended not to resize the screen when you have pressed the "Numbers Trivia" button. If it has happened, and the website UI looks weird. Please refresh the page again